pyproject.toml
[tool.poetry]
name = "pdninja"
version = "0.1.0"
description = ""
authors = ["Shubham Agarwal <shubhamagarwal27121996@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pysigma = "^0.11.14"
pysigma-backend-pd-df = "^0.1.0"
pysigma-pipeline-sysmon = "^1.0.4"
pysigma-pipeline-windows = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
